requirements.txt
discord.py==2.0.0
pytz==2021.3
